refactor: move functions to guards file
